refactor: extracted WorkoutRunner's partial widgets into own files
